feat:FSM orchestration module definition

+++ VerticalWarehouses/Common/Ferretto.VW.CommonUtils/Messages/Data/WeightAcquisitionCommandMessageData.cs
@@ -6,17 +6,19 @@ namespace Ferretto.VW.CommonUtils.Messages.Data
 {
     public class WeightAcquisitionCommandMessageData : IMessageData
     {
+
+
         #region Properties
 
         public CommandAction CommandAction { get; set; }
 
-        public double Displacement { get; set; }
+        public decimal Displacement { get; set; }
 
-        public double InitialPosition { get; set; }
+        public decimal InitialPosition { get; set; }
 
         public TimeSpan InPlaceSamplingDuration { get; set; }
 
-        public double NetWeight { get; set; }
+        public decimal NetWeight { get; set; }
 
         public MessageVerbosity Verbosity { get; set; }
