+++ VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.MissionsManager/Providers/IWeightAnalysisMissionProvider.cs
@@ -2,9 +2,11 @@
 {
     public interface IWeightAnalysisMissionProvider
     {
+
+
         #region Methods
 
-        void Start(double initialPosition, double displacement, double netWeight, System.TimeSpan inPlaceSamplingDuration);
+        void Start(decimal initialPosition, decimal displacement, decimal netWeight, System.TimeSpan inPlaceSamplingDuration);
 
         void Stop();
